+++ Games/Oxide.TheForest/TheForestPlugin.cs
@@ -1,0 +1,6 @@
+﻿namespace Oxide.Plugins
+{
+    public abstract class TheForestPlugin : CSharpPlugin
+    {
+    }
+}
