--- Games/Unity/Oxide.Game.TheForest/TheForest.csproj
@@ -1,135 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="12.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <Import Project="$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props" Condition="Exists('$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props')" />
-  <PropertyGroup>
-    <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
-    <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{183056B0-1635-46BB-81CC-8669EE46E83F}</ProjectGuid>
-    <OutputType>Library</OutputType>
-    <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>Oxide.Game.TheForest</RootNamespace>
-    <AssemblyName>Oxide.Game.TheForest</AssemblyName>
-    <GameName>$(MSBuildProjectName)</GameName>
-    <BundleDir>$(SolutionDir)Bundles</BundleDir>
-    <ManagedDir>TheForestDedicatedServer_Data\Managed</ManagedDir>
-    <TargetFrameworkVersion>v3.5</TargetFrameworkVersion>
-    <FileAlignment>512</FileAlignment>
-    <TargetFrameworkProfile />
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
-    <DebugSymbols>true</DebugSymbols>
-    <OutputPath>bin\Debug\</OutputPath>
-    <DefineConstants>DEBUG;TRACE</DefineConstants>
-    <DebugType>full</DebugType>
-    <PlatformTarget>AnyCPU</PlatformTarget>
-    <ErrorReport>prompt</ErrorReport>
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
-    <OutputPath>bin\Release\</OutputPath>
-    <DefineConstants>TRACE</DefineConstants>
-    <Optimize>true</Optimize>
-    <DebugType>pdbonly</DebugType>
-    <PlatformTarget>AnyCPU</PlatformTarget>
-    <ErrorReport>prompt</ErrorReport>
-  </PropertyGroup>
-  <ItemGroup>
-    <Reference Include="Oxide.References">
-      <HintPath>..\..\..\Oxide.References\bin\$(Configuration)\net35\Merged\Oxide.References.dll</HintPath>
-    </Reference>
-    <Reference Include="Assembly-CSharp">
-      <HintPath>Patched\Assembly-CSharp.dll</HintPath>
-    </Reference>
-    <Reference Include="Assembly-CSharp-firstpass">
-      <HintPath>Dependencies\Assembly-CSharp-firstpass.dll</HintPath>
-    </Reference>
-    <Reference Include="bolt">
-      <HintPath>Dependencies\bolt.dll</HintPath>
-    </Reference>
-    <Reference Include="bolt.user">
-      <HintPath>Dependencies\bolt.user.dll</HintPath>
-    </Reference>
-    <Reference Include="PlayMaker">
-      <HintPath>Dependencies\PlayMaker.dll</HintPath>
-    </Reference>
-    <Reference Include="System" />
-    <Reference Include="System.Core" />
-    <Reference Include="udpkit">
-      <HintPath>Dependencies\udpkit.dll</HintPath>
-    </Reference>
-    <Reference Include="udpkit.common">
-      <HintPath>Dependencies\udpkit.common.dll</HintPath>
-    </Reference>
-    <Reference Include="UnityEngine">
-      <HintPath>Dependencies\UnityEngine.dll</HintPath>
-    </Reference>
-  </ItemGroup>
-  <ItemGroup>
-    <Compile Include="Libraries\Covalence\TheForestCommandSystem.cs" />
-    <Compile Include="Libraries\Covalence\TheForestConsolePlayer.cs" />
-    <Compile Include="Libraries\Covalence\TheForestCovalenceProvider.cs" />
-    <Compile Include="Libraries\Covalence\TheForestPlayer.cs" />
-    <Compile Include="Libraries\Covalence\TheForestPlayerManager.cs" />
-    <Compile Include="Libraries\Covalence\TheForestServer.cs" />
-    <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="TheForestCommands.cs" />
-    <Compile Include="TheForestCore.cs" />
-    <Compile Include="TheForestExtension.cs" />
-    <Compile Include="TheForestHooks.cs" />
-    <Compile Include="TheForestPlugin.cs" />
-    <Compile Include="TheForestPluginLoader.cs" />
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\..\..\Oxide.Core\Core.csproj">
-      <Project>{67f65f5d-a62a-4752-849d-bff30d828107}</Project>
-      <Name>Core</Name>
-    </ProjectReference>
-    <ProjectReference Include="..\..\..\Extensions\Oxide.CSharp\CSharp.csproj">
-      <Project>{9103d682-d1aa-4a95-a499-896f551aaa62}</Project>
-      <Name>CSharp</Name>
-    </ProjectReference>
-  </ItemGroup>
-  <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
-  <Target Name="Clean">
-    <RemoveDir Directories="$(BundleDir)\$(GameName)" />
-  </Target>
-  <Target Name="ChangeAliasesOfStrongNameAssemblies" BeforeTargets="FindReferenceAssembliesForReferences;ResolveReferences">
-    <ItemGroup>
-      <ReferencePath Condition="'%(FileName)' == 'Oxide.References'">
-        <Aliases>Oxide</Aliases>
-      </ReferencePath>
-    </ItemGroup>
-  </Target>
-  <Target Name="AfterBuild">
-    <ItemGroup>
-      <GameFiles Include="$(ProjectDir)Files\**\*.*" />
-      <ManagedFiles Include="$(SolutionDir)Extensions\Oxide.CSharp\bin\$(Configuration)\net35\Oxide.CSharp.dll" />
-      <ManagedFiles Include="$(SolutionDir)Extensions\Oxide.CSharp\Dependencies\**\*.*" />
-      <ManagedFiles Include="$(SolutionDir)Extensions\Oxide.MySql\bin\$(Configuration)\net35\Oxide.MySql.dll" />
-      <ManagedFiles Include="$(SolutionDir)Extensions\Oxide.MySql\Dependencies\**\*.*" />
-      <ManagedFiles Include="$(SolutionDir)Extensions\Oxide.SQLite\bin\$(Configuration)\net35\Oxide.SQLite.dll" />
-      <ManagedFiles Include="$(SolutionDir)Extensions\Oxide.SQLite\Dependencies\**\*.*" />
-      <ManagedFiles Include="$(SolutionDir)Extensions\Oxide.Unity\bin\$(Configuration)\net35\Oxide.Unity.dll" />
-      <ManagedFiles Include="$(SolutionDir)Oxide.Core\bin\$(Configuration)\net35\Oxide.*.dll" />
-      <ManagedFiles Include="$(SolutionDir)Oxide.References\Dependencies\**\*.*" Exclude="$(SolutionDir)Oxide.References\Dependencies\Mono.*.dll" />
-      <ManagedFiles Include="$(TargetDir)$(AssemblyName).dll" />
-      <PatchedFiles Include="$(ProjectDir)Patched\**\*.*" />
-    </ItemGroup>
-    <Exec Command="pdb2mdb %(ManagedFiles.Identity)" Condition="'$(Configuration)' == 'Debug' And '$(OS)' == 'Windows_NT'" IgnoreExitCode="true" />
-    <Copy SourceFiles="@(GameFiles)" DestinationFiles="@(GameFiles->'$(BundleDir)\$(GameName)\%(RecursiveDir)%(Filename)%(Extension)')" />
-    <Copy SourceFiles="@(ManagedFiles)" DestinationFiles="@(ManagedFiles->'$(BundleDir)\$(GameName)\$(ManagedDir)\%(RecursiveDir)%(Filename)%(Extension)')" />
-    <Copy SourceFiles="@(PackageFiles)" DestinationFiles="@(PackageFiles->'$(BundleDir)\$(GameName)\$(ManagedDir)\%(Filename)%(Extension)')" />
-    <Copy SourceFiles="@(PatchedFiles)" DestinationFiles="@(PatchedFiles->'$(BundleDir)\$(GameName)\$(ManagedDir)\%(RecursiveDir)%(Filename)%(Extension)')" />
-    <CallTarget Targets="Deploy" Condition="Exists('$(ProjectDir).deploy')" />
-  </Target>
-  <Target Name="Deploy">
-    <Exec Command="start /b /wait taskkill /f /im TheForest.exe /fi &quot;memusage gt 2&quot;" Condition="'$(OS)' == 'Windows_NT'" />
-    <Exec Command="start /b /wait taskkill /f /im CSharpCompiler.exe /fi &quot;memusage gt 2&quot;" Condition="'$(OS)' == 'Windows_NT'" />
-    <ReadLinesFromFile File="$(ProjectDir).deploy">
-      <Output PropertyName="DeployPath" TaskParameter="Lines" />
-    </ReadLinesFromFile>
-    <ItemGroup>
-      <DeployFiles Include="$(BundleDir)\$(GameName)\**\*.*" />
-    </ItemGroup>
-    <Copy Condition="'$(DeployPath)' != ''" SourceFiles="@(DeployFiles)" DestinationFiles="@(DeployFiles->'$(DeployPath)\%(RecursiveDir)%(Filename)%(Extension)')" />
-  </Target>
-</Project>
