--- Games/Unity/Oxide.Game.TheForest/TheForestCore.cs
@@ -1,185 +1,0 @@
-﻿﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using Oxide.Core;
-using Oxide.Core.Libraries;
-using Oxide.Core.Libraries.Covalence;
-using Oxide.Core.Plugins;
-using Oxide.Game.TheForest.Libraries.Covalence;
-using Steamworks;
-using TheForest.Utils;
-
-namespace Oxide.Game.TheForest
-{
-    /// <summary>
-    /// The core The Forest plugin
-    /// </summary>
-    public partial class TheForestCore : CSPlugin
-    {
-        #region Initialization
-
-        /// <summary>
-        /// Initializes a new instance of the TheForestCore class
-        /// </summary>
-        public TheForestCore()
-        {
-            // Set plugin info attributes
-            Title = "The Forest";
-            Author = TheForestExtension.AssemblyAuthors;
-            Version = TheForestExtension.AssemblyVersion;
-        }
-
-        // Libraries
-        //internal readonly Command cmdlib = Interface.Oxide.GetLibrary<Command>();
-        internal readonly Lang lang = Interface.Oxide.GetLibrary<Lang>();
-        internal readonly Permission permission = Interface.Oxide.GetLibrary<Permission>();
-        //internal readonly Player Player = Interface.Oxide.GetLibrary<Player>();
-
-        // Instances
-        internal static readonly TheForestCovalenceProvider Covalence = TheForestCovalenceProvider.Instance;
-        internal readonly PluginManager pluginManager = Interface.Oxide.RootPluginManager;
-        internal readonly IServer Server = Covalence.CreateServer();
-
-        // Commands that a plugin can't override
-        internal static IEnumerable<string> RestrictedCommands => new[]
-        {
-            ""
-        };
-
-        private bool serverInitialized;
-
-        #endregion
-
-        #region Core Hooks
-
-        [HookMethod("Init")]
-        private void Init()
-        {
-            // Configure remote error logging
-            RemoteLogger.SetTag("game", Title.ToLower());
-            RemoteLogger.SetTag("game version", Server.Version);
-
-            // Add core plugin commands
-            AddCovalenceCommand(new[] { "oxide.plugins", "o.plugins", "plugins" }, "PluginsCommand", "oxide.plugins");
-            AddCovalenceCommand(new[] { "oxide.load", "o.load", "plugin.load" }, "LoadCommand", "oxide.load");
-            AddCovalenceCommand(new[] { "oxide.reload", "o.reload", "plugin.reload" }, "ReloadCommand", "oxide.reload");
-            AddCovalenceCommand(new[] { "oxide.unload", "o.unload", "plugin.unload" }, "UnloadCommand", "oxide.unload");
-
-            // Add core permission commands
-            AddCovalenceCommand(new[] { "oxide.grant", "o.grant", "perm.grant" }, "GrantCommand", "oxide.grant");
-            AddCovalenceCommand(new[] { "oxide.group", "o.group", "perm.group" }, "GroupCommand", "oxide.group");
-            AddCovalenceCommand(new[] { "oxide.revoke", "o.revoke", "perm.revoke" }, "RevokeCommand", "oxide.revoke");
-            AddCovalenceCommand(new[] { "oxide.show", "o.show", "perm.show" }, "ShowCommand", "oxide.show");
-            AddCovalenceCommand(new[] { "oxide.usergroup", "o.usergroup", "perm.usergroup" }, "UserGroupCommand", "oxide.usergroup");
-
-            // Add core misc commands
-            AddCovalenceCommand(new[] { "oxide.lang", "o.lang" }, "LangCommand");
-            AddCovalenceCommand(new[] { "oxide.version", "o.version" }, "VersionCommand");
-
-            // Register messages for localization
-            foreach (var language in Core.Localization.languages) lang.RegisterMessages(language.Value, this, language.Key);
-
-            // Setup default permission groups
-            if (permission.IsLoaded)
-            {
-                var rank = 0;
-                foreach (var defaultGroup in Interface.Oxide.Config.Options.DefaultGroups)
-                    if (!permission.GroupExists(defaultGroup)) permission.CreateGroup(defaultGroup, defaultGroup, rank++);
-
-                permission.RegisterValidate(s =>
-                {
-                    ulong temp;
-                    if (!ulong.TryParse(s, out temp)) return false;
-                    var digits = temp == 0 ? 1 : (int)Math.Floor(Math.Log10(temp) + 1);
-                    return digits >= 17;
-                });
-
-                permission.CleanUp();
-            }
-        }
-
-        [HookMethod("OnPluginLoaded")]
-        private void OnPluginLoaded(Plugin plugin)
-        {
-            // Call OnServerInitialized for hotloaded plugins
-            if (serverInitialized) plugin.CallHook("OnServerInitialized");
-        }
-
-        [HookMethod("OnServerInitialized")]
-        private void OnServerInitialized()
-        {
-            if (serverInitialized) return;
-
-            Analytics.Collect();
-            TheForestExtension.ServerConsole();
-            SteamGameServer.SetGameTags("oxide,modded");
-
-            serverInitialized = true;
-        }
-
-        [HookMethod("OnServerShutdown")]
-        private void OnServerShutdown() => Interface.Oxide.OnShutdown();
-
-        #endregion
-
-        #region Command Handling
-
-        [HookMethod("IOnServerCommand")]
-        private object IOnServerCommand(BoltConnection connection, string command, string data)
-        {
-            if (command.Length == 0) return null;
-
-            // Get the full command
-            var cmd = command.TrimStart('/');
-            var args = string.IsNullOrEmpty(data) ? new string[] {} : data.Split();
-
-            if (Interface.Call("OnServerCommand", cmd, args) != null) return true;
-
-            // Check if command is from the player
-            if (connection == null) return null;
-
-            var id = connection.RemoteEndPoint.SteamId.Id.ToString();
-            var entity = Scene.SceneTracker.allPlayerEntities.FirstOrDefault(ent => ent.source.ConnectionId == connection.ConnectionId);
-            if (entity == null) return null;
-
-            // Get the covalence player
-            var iplayer = Covalence.PlayerManager.FindPlayerById(id);
-            if (iplayer == null) return null;
-
-            // Is the command blocked?
-            var blockedSpecific = Interface.Call("OnPlayerCommand", entity, cmd, args);
-            var blockedCovalence = Interface.Call("OnUserCommand", iplayer, cmd, args);
-            if (blockedSpecific != null || blockedCovalence != null) return true;
-
-            // Is it a chat command?
-            if (command[0] != '/') return null;
-
-            // Is it a covalance command?
-            if (Covalence.CommandSystem.HandleChatMessage(iplayer, command)) return true;
-
-            // Is it a regular chat command?
-            //if (!cmdlib.HandleChatCommand(player, cmd, args)) // TODO: Implement
-            //    iplayer.Reply(lang.GetMessage("UnknownCommand", this, iplayer.Id), cmd);
-
-            return true;
-        }
-
-        #endregion
-
-        #region Helpers
-
-        /// <summary>
-        /// Checks if the permission system has loaded, shows an error if it failed to load
-        /// </summary>
-        /// <param name="player"></param>
-        /// <returns></returns>
-        private bool PermissionsLoaded(IPlayer player)
-        {
-            if (permission.IsLoaded) return true;
-            player.Reply(lang.GetMessage("PermissionsNotLoaded", this, player.Id), permission.LastException.Message);
-            return false;
-        }
-
-        #endregion
-    }
-}
