--- Games/Unity/Oxide.Game.TheForest/TheForestPlugin.cs
@@ -1,6 +1,0 @@
-﻿namespace Oxide.Plugins
-{
-    public abstract class TheForestPlugin : CSharpPlugin
-    {
-    }
-}
