+++ Oxide.sln
@@ -188,7 +188,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Terraria", "Games\Other\Oxi
 		{7009BBCB-8CA4-49EC-BE95-B27C07771CDF} = {7009BBCB-8CA4-49EC-BE95-B27C07771CDF}
 	EndProjectSection
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TheForest", "Games\Unity\Oxide.Game.TheForest\TheForest.csproj", "{183056B0-1635-46BB-81CC-8669EE46E83F}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TheForest", "Games\Oxide.TheForest\TheForest.csproj", "{183056B0-1635-46BB-81CC-8669EE46E83F}"
 	ProjectSection(ProjectDependencies) = postProject
 		{C54EA613-C5FD-4ACD-8AC4-79063B08CF5A} = {C54EA613-C5FD-4ACD-8AC4-79063B08CF5A}
 		{D138493B-6F2E-4B9F-8F79-17B73422D87F} = {D138493B-6F2E-4B9F-8F79-17B73422D87F}
